Conflicts:
	AUTHORS
	Makefile.flags
	coreutils/Config.in
	coreutils/Kbuild
	coreutils/date.c
	coreutils/test.c
	include/applets.h
	include/libbb.h
	include/usage.h
	libbb/getopt32.c
	libbb/xconnect.c
	loginutils/passwd.c
	miscutils/Config.in
	miscutils/Kbuild
	miscutils/perd.c
	networking/Config.in
	networking/Kbuild
	networking/nslookup.c
	networking/ping.c
	networking/telnetd.c
	networking/tftp.c
	networking/traceroute.c
	shell/ash.c
	shell/hush.c
	util-linux/Config.in

+++ procps/kill.c
@@ -5,8 +5,76 @@
  * Copyright (C) 1995, 1996 by Bruce Perens <bruce@pixar.com>.
  * Copyright (C) 1999-2004 by Erik Andersen <andersen@codepoet.org>
  *
- * Licensed under the GPL v2 or later, see the file LICENSE in this tarball.
+ * Licensed under GPLv2 or later, see file LICENSE in this source tree.
  */
+//config:config KILL
+//config:	bool "kill"
+//config:	default y
+//config:	help
+//config:	  The command kill sends the specified signal to the specified
+//config:	  process or process group. If no signal is specified, the TERM
+//config:	  signal is sent.
+//config:
+//config:config KILLALL
+//config:	bool "killall"
+//config:	default y
+//config:	help
+//config:	  killall sends a signal to all processes running any of the
+//config:	  specified commands. If no signal name is specified, SIGTERM is
+//config:	  sent.
+//config:
+//config:config KILLALL5
+//config:	bool "killall5"
+//config:	default y
+//config:	help
+//config:	  The SystemV killall command. killall5 sends a signal
+//config:	  to all processes except kernel threads and the processes
+//config:	  in its own session, so it won't kill the shell that is running
+//config:	  the script it was called from.
+
+//applet:IF_KILL(APPLET(kill, BB_DIR_BIN, BB_SUID_DROP))
+//applet:IF_KILLALL(APPLET_ODDNAME(killall, kill, BB_DIR_USR_BIN, BB_SUID_DROP, killall))
+//applet:IF_KILLALL5(APPLET_ODDNAME(killall5, kill, BB_DIR_USR_SBIN, BB_SUID_DROP, killall5))
+
+//kbuild:lib-$(CONFIG_KILL) += kill.o
+//kbuild:lib-$(CONFIG_KILLALL) += kill.o
+//kbuild:lib-$(CONFIG_KILLALL5) += kill.o
+
+//usage:#define kill_trivial_usage
+//usage:       "[-l] [-SIG] PID..."
+//usage:#define kill_full_usage "\n\n"
+//usage:       "Send a signal (default: TERM) to given PIDs\n"
+//usage:     "\n	-l	List all signal names and numbers"
+/* //usage:  "\n	-s SIG	Yet another way of specifying SIG" */
+//usage:
+//usage:#define kill_example_usage
+//usage:       "$ ps | grep apache\n"
+//usage:       "252 root     root     S [apache]\n"
+//usage:       "263 www-data www-data S [apache]\n"
+//usage:       "264 www-data www-data S [apache]\n"
+//usage:       "265 www-data www-data S [apache]\n"
+//usage:       "266 www-data www-data S [apache]\n"
+//usage:       "267 www-data www-data S [apache]\n"
+//usage:       "$ kill 252\n"
+//usage:
+//usage:#define killall_trivial_usage
+//usage:       "[-l] [-q] [-SIG] PROCESS_NAME..."
+//usage:#define killall_full_usage "\n\n"
+//usage:       "Send a signal (default: TERM) to given processes\n"
+//usage:     "\n	-l	List all signal names and numbers"
+/* //usage:  "\n	-s SIG	Yet another way of specifying SIG" */
+//usage:     "\n	-q	Don't complain if no processes were killed"
+//usage:
+//usage:#define killall_example_usage
+//usage:       "$ killall apache\n"
+//usage:
+//usage:#define killall5_trivial_usage
+//usage:       "[-l] [-SIG] [-o PID]..."
+//usage:#define killall5_full_usage "\n\n"
+//usage:       "Send a signal (default: TERM) to all processes outside current session\n"
+//usage:     "\n	-l	List all signal names and numbers"
+//usage:     "\n	-o PID	Don't signal this PID"
+/* //usage:  "\n	-s SIG	Yet another way of specifying SIG" */
 
 #include "libbb.h"
 
@@ -24,29 +92,34 @@
  * This is needed to avoid collision with kill -9 ... syntax
  */
 
-int kill_main(int argc, char **argv)
+int kill_main(int argc UNUSED_PARAM, char **argv)
 {
 	char *arg;
 	pid_t pid;
 	int signo = SIGTERM, errors = 0, quiet = 0;
-#if !ENABLE_KILLALL && !ENABLE_KILLALL5
-#define killall 0
-#define killall5 0
+#if ENABLE_KILL && !ENABLE_KILLALL && !ENABLE_KILLALL5
+# define killall  0
+# define killall5 0
+#elif !ENABLE_KILL && ENABLE_KILLALL && !ENABLE_KILLALL5
+# define killall  1
+# define killall5 0
+#elif !ENABLE_KILL && !ENABLE_KILLALL && ENABLE_KILLALL5
+# define killall  0
+# define killall5 1
 #else
 /* How to determine who we are? find 3rd char from the end:
  * kill, killall, killall5
  *  ^i       ^a        ^l  - it's unique
  * (checking from the start is complicated by /bin/kill... case) */
 	const char char3 = argv[0][strlen(argv[0]) - 3];
-#define killall (ENABLE_KILLALL && char3 == 'a')
-#define killall5 (ENABLE_KILLALL5 && char3 == 'l')
+# define killall  (ENABLE_KILLALL && char3 == 'a')
+# define killall5 (ENABLE_KILLALL5 && char3 == 'l')
 #endif
 
 	/* Parse any options */
-	argc--;
 	arg = *++argv;
 
-	if (argc < 1 || arg[0] != '-') {
+	if (!arg || arg[0] != '-') {
 		goto do_it_now;
 	}
 
@@ -55,13 +128,14 @@ int kill_main(int argc, char **argv)
 	 * echo "Died of SIG`kill -l $?`"
 	 * We try to mimic what kill from coreutils-6.8 does */
 	if (arg[1] == 'l' && arg[2] == '\0') {
-		if (argc == 1) {
+		arg = *++argv;
+		if (!arg) {
 			/* Print the whole signal list */
 			print_signames();
 			return 0;
 		}
 		/* -l <sig list> */
-		while ((arg = *++argv)) {
+		do {
 			if (isdigit(arg[0])) {
 				signo = bb_strtou(arg, NULL, 10);
 				if (errno) {
@@ -82,8 +156,8 @@ int kill_main(int argc, char **argv)
 				}
 				printf("%d\n", signo);
 			}
-		}
-		/* If they specified -l, we are all done */
+			arg = *++argv;
+		} while (arg);
 		return EXIT_SUCCESS;
 	}
 
@@ -91,14 +165,19 @@ int kill_main(int argc, char **argv)
 	if (killall && arg[1] == 'q' && arg[2] == '\0') {
 		quiet = 1;
 		arg = *++argv;
-		argc--;
-		if (argc < 1) bb_show_usage();
-		if (arg[0] != '-') goto do_it_now;
+		if (!arg)
+			bb_show_usage();
+		if (arg[0] != '-')
+			goto do_it_now;
 	}
 
 	arg++; /* skip '-' */
-	if (argc > 1 && arg[0] == 's' && arg[1] == '\0') { /* -s SIG? */
-		argc--;
+
+	/* -o PID? (if present, it always is at the end of command line) */
+	if (killall5 && arg[0] == 'o')
+		goto do_it_now;
+
+	if (argv[1] && arg[0] == 's' && arg[1] == '\0') { /* -s SIG? */
 		arg = *++argv;
 	} /* else it must be -SIG */
 	signo = get_signum(arg);
@@ -107,38 +186,77 @@ int kill_main(int argc, char **argv)
 		return EXIT_FAILURE;
 	}
 	arg = *++argv;
-	argc--;
 
-do_it_now:
+ do_it_now:
 	pid = getpid();
 
 	if (killall5) {
 		pid_t sid;
 		procps_status_t* p = NULL;
+		/* compat: exitcode 2 is "no one was signaled" */
+		int ret = 2;
 
-		/* Find out our own session id */
+		/* Find out our session id */
 		sid = getsid(pid);
-		/* Now stop all processes */
-		kill(-1, SIGSTOP);
-		/* Now kill all processes except our session */
-		while ((p = procps_scan(p, PSSCAN_PID|PSSCAN_SID))) {
-			if (p->sid != (unsigned)sid && p->pid != (unsigned)pid && p->pid != 1)
-				kill(p->pid, signo);
+		/* Stop all processes */
+		if (signo != SIGSTOP && signo != SIGCONT)
+			kill(-1, SIGSTOP);
+		/* Signal all processes except those in our session */
+		while ((p = procps_scan(p, PSSCAN_PID|PSSCAN_SID)) != NULL) {
+			char **args;
+
+			if (p->sid == (unsigned)sid
+			 || p->sid == 0 /* compat: kernel thread, don't signal it */
+			 || p->pid == (unsigned)pid
+			 || p->pid == 1
+			) {
+				continue;
+			}
+
+			/* All remaining args must be -o PID options.
+			 * Check p->pid against them. */
+			args = argv;
+			while (*args) {
+				pid_t omit;
+
+				arg = *args++;
+				if (arg[0] != '-' || arg[1] != 'o') {
+					bb_error_msg("bad option '%s'", arg);
+					ret = 1;
+					goto resume;
+				}
+				arg += 2;
+				if (!arg[0] && *args)
+					arg = *args++;
+				omit = bb_strtoi(arg, NULL, 10);
+				if (errno) {
+					bb_error_msg("invalid number '%s'", arg);
+					ret = 1;
+					goto resume;
+				}
+				if (p->pid == omit)
+					goto dont_kill;
+			}
+			kill(p->pid, signo);
+			ret = 0;
+ dont_kill: ;
 		}
+ resume:
 		/* And let them continue */
-		kill(-1, SIGCONT);
-		return 0;
+		if (signo != SIGSTOP && signo != SIGCONT)
+			kill(-1, SIGCONT);
+		return ret;
 	}
 
 	/* Pid or name is required for kill/killall */
-	if (argc < 1) {
+	if (!arg) {
 		bb_error_msg("you need to specify whom to kill");
 		return EXIT_FAILURE;
 	}
 
 	if (killall) {
 		/* Looks like they want to do a killall.  Do that */
-		while (arg) {
+		do {
 			pid_t* pidList;
 
 			pidList = find_pid_by_name(arg);
@@ -156,28 +274,49 @@ do_it_now:
 						continue;
 					errors++;
 					if (!quiet)
-						bb_perror_msg("cannot kill pid %u", (unsigned)*pl);
+						bb_perror_msg("can't kill pid %d", (int)*pl);
 				}
 			}
 			free(pidList);
 			arg = *++argv;
-		}
+		} while (arg);
 		return errors;
 	}
 
 	/* Looks like they want to do a kill. Do that */
 	while (arg) {
-		/* Support shell 'space' trick */
-		if (arg[0] == ' ')
-			arg++;
+#if ENABLE_ASH || ENABLE_HUSH
+		/*
+		 * We need to support shell's "hack formats" of
+		 * " -PRGP_ID" (yes, with a leading space)
+		 * and " PID1 PID2 PID3" (with degenerate case "")
+		 */
+		while (*arg != '\0') {
+			char *end;
+			if (*arg == ' ')
+				arg++;
+			pid = bb_strtoi(arg, &end, 10);
+			if (errno && (errno != EINVAL || *end != ' ')) {
+				bb_error_msg("invalid number '%s'", arg);
+				errors++;
+				break;
+			}
+			if (kill(pid, signo) != 0) {
+				bb_perror_msg("can't kill pid %d", (int)pid);
+				errors++;
+			}
+			arg = end; /* can only point to ' ' or '\0' now */
+		}
+#else
 		pid = bb_strtoi(arg, NULL, 10);
 		if (errno) {
-			bb_error_msg("bad pid '%s'", arg);
+			bb_error_msg("invalid number '%s'", arg);
 			errors++;
 		} else if (kill(pid, signo) != 0) {
-			bb_perror_msg("cannot kill pid %d", (int)pid);
+			bb_perror_msg("can't kill pid %d", (int)pid);
 			errors++;
 		}
+#endif
 		arg = *++argv;
 	}
 	return errors;
